requirements.txt
asgiref==3.7.2
asttokens==2.2.1
Babel==2.12.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-autoslug==1.9.9
django-ckeditor==6.6.1
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-werkzeug-debugger-runserver==0.3.1
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
phonenumbers==8.13.15
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
